moved auth files to their own directory
